finish

+++ controllers/restaurantController.ts
@@ -1,7 +1,7 @@
 import { Request, Response, query } from 'express'
 import db from '../db/db'
 
-type QueryConfig = {
+export type QueryConfig = {
   text: string
   values: any[]
 }
@@ -28,12 +28,28 @@ class restaurantController {
   }
 
   getRestaurant = async (req: Request, res: Response) => {
-    const query: QueryConfig = {
+    // get restaurant by id
+    const query1: QueryConfig = {
       text: 'SELECT * FROM restaurant WHERE id = $1',
       values: [req.params.id],
     }
-    const restaurant = await db.query(query)
-    res.status(200).json(restaurant.rows)
+    const restaurant = await db.query(query1)
+    if (restaurant.rows.length === 0) {
+      res.status(404).json({
+        message: 'Restaurant not found',
+      })
+      return
+    }
+    // get all dishes by restaurant id
+    const query2: QueryConfig = {
+      text: 'SELECT * FROM dish WHERE restaurantId = $1',
+      values: [req.params.id],
+    }
+    const dishes = await db.query(query2)
+    res.status(200).json({
+      ...restaurant.rows[0],
+      dishes: dishes.rows,
+    })
   }
   createRestaurant = async (req: Request, res: Response) => {
     const query: QueryConfig = {
@@ -41,10 +57,22 @@ class restaurantController {
       values: [req.body.name, req.body.isKosher, 0, req.body.cuisines],
     }
     await db.query(query)
-    res.status(201).send('Restaurant Created')
+    res.sendStatus(201)
   }
 
   updateRestaurant = async (req: Request, res: Response) => {
+    // check if restaurant exists
+    const restaurantQuery: QueryConfig = {
+      text: 'SELECT * FROM restaurant WHERE id = $1',
+      values: [req.params.id],
+    }
+    const restaurant = await db.query(restaurantQuery)
+    if (restaurant.rows.length === 0) {
+      res.status(404).json({
+        message: 'Restaurant not found',
+      })
+      return
+    }
     let queryString = 'UPDATE restaurant SET '
     const values = [req.params.id]
     const setClauses = []
@@ -62,16 +90,41 @@ class restaurantController {
 
     queryString += setClauses.join(', ') + ' WHERE id = $1'
     await db.query(queryString, values)
-    res.status(200).send('Restaurant Updated')
+    res.sendStatus(200)
   }
 
   deleteRestaurant = async (req: Request, res: Response) => {
+    // check if restaurant exists
+    const restaurantQuery: QueryConfig = {
+      text: 'SELECT * FROM restaurant WHERE id = $1',
+      values: [req.params.id],
+    }
+    const restaurant = await db.query(restaurantQuery)
+    if (restaurant.rows.length === 0) {
+      res.status(404).json({
+        message: 'Restaurant not found',
+      })
+      return
+    }
+
     const query: QueryConfig = {
       text: 'DELETE FROM restaurant WHERE id = $1',
       values: [req.params.id],
     }
+
+    // delete orders of this restaurant
+    await db.query('DELETE FROM orders WHERE restaurantId = $1', [
+      req.params.id,
+    ])
+    // delete dishes of this restaurant
+    await db.query('DELETE FROM dish WHERE restaurantId = $1', [req.params.id])
+    // delete ratings of this restaurant
+    await db.query('DELETE FROM rating WHERE restaurantId = $1', [
+      req.params.id,
+    ])
+
     await db.query(query)
-    res.status(204).send('Restaurant Deleted')
+    res.sendStatus(204)
   }
 }
 export default new restaurantController()
